requirements.txt
Django==1.5.1
MySQL-python
distribute==0.6.31
django-bootstrap-toolkit==2.12.0
wsgiref==0.1.2
django-endless-pagination
South
django-tastypie
